+++ app/Http/Controllers/AngsuranPenjualanController.php
@@ -649,11 +649,14 @@ class AngsuranPenjualanController extends Controller
             $transaksi->jumlah_pembayaran=$jumlahpembayaran;
             $transaksi->save();
 
-        if ($tableangsuran->delete())
-        {
-            $isi=Auth::user()->username." telah menghapus angsuran penjualan dengan No. Angsuran ".$tableangsuran->id." pada No. Transaksi Penjualan ".$idtrans." di Cabang ".Auth::user()->cabangs->Nama_Cabang.".";
-            $save=$this->createlog($isi,"delete");
-            return "{\"msg\":\"success\"}";
+        $tableangsuran->reason_on_delete = $request->json('reason_on_delete');
+        if ($tableangsuran->save()) {
+          if ($tableangsuran->delete())
+          {
+              $isi=Auth::user()->username." telah menghapus angsuran penjualan dengan No. Angsuran ".$tableangsuran->id." pada No. Transaksi Penjualan ".$idtrans." di Cabang ".Auth::user()->cabangs->Nama_Cabang.".";
+              $save=$this->createlog($isi,"delete");
+              return "{\"msg\":\"success\"}";
+          }
         }
         else
         {
